+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`__.
 ------------
 
 - Removed kicked-user-made-bot-think-he-was-kicked bug
+- Using generic RabbitMQ wrapper for easy reuse
 
 [1.0] - 2016-04-14
 ------------------